Updated component styles, relocated product model and service files.
