+++ src/app/app.component.css
@@ -3,6 +3,3 @@ h1 {
   font-family: Arial, Helvetica, sans-serif;
   font-size: 250%;
 }
-#container {
-  margin: 2em;
-}
