+++ src/app/app.component.html
@@ -1,4 +1,4 @@
-<div id="container">
+<div class="container">
   <h1>
     {{title}}
   </h1>
