+++ src/app/products/products.component.html
@@ -13,6 +13,7 @@
                 </thead>
                 <tbody>
                     <tr *ngFor="let product of products">
+                        <td>{{ product.productId }}</td>
                         <td>{{ product.productName }}</td>
                         <td>{{ product.unitPrice }}</td>
                     </tr>
