+++ src/app/shared/models/product.ts
